Simplify session log source planning
