+++ src/meridian/lib/ops/session_target.py
@@ -40,19 +40,20 @@ _PRIMARY_TRANSCRIPT_UNAVAILABLE_SUFFIX = (
 )
 
 
+class TranscriptSource(NamedTuple):
+    kind: Literal["file", "opencode_db", "spawn_history"]
+    session_id: str
+    harness: str | None
+    source_label: str
+    path: Path | None = None
+
+
 class SessionLogTarget(NamedTuple):
     session_id: str
     harness: str | None
     file_path: Path | None
     source: str
-    transcript_source: Literal["file", "opencode_db"] = "file"
-    fallback_targets: tuple[SessionLogTarget, ...] = ()
-
-
-class SessionRepairTarget(NamedTuple):
-    detected_harness_session_id: str | None
-    source: str | None
-    reason: str | None = None
+    sources: tuple[TranscriptSource, ...]
 
 
 def _is_chat_ref(runtime_root: Path, value: str) -> bool:
@@ -72,6 +73,43 @@ def _extract_session_id_from_path(path: Path) -> str:
     return path.name
 
 
+def _target_from_source(source: TranscriptSource) -> SessionLogTarget:
+    return SessionLogTarget(
+        session_id=source.session_id,
+        harness=source.harness,
+        file_path=source.path,
+        source=source.source_label,
+        sources=(source,),
+    )
+
+
+def _source_key(source: TranscriptSource) -> tuple[str, str, str | None, str]:
+    return (
+        source.kind,
+        source.session_id,
+        source.path.as_posix() if source.path is not None else None,
+        source.source_label,
+    )
+
+
+def _with_sources(
+    target: SessionLogTarget,
+    *additional_targets: SessionLogTarget | None,
+) -> SessionLogTarget:
+    sources = list(target.sources)
+    seen = {_source_key(source) for source in sources}
+    for additional_target in additional_targets:
+        if additional_target is None:
+            continue
+        for source in additional_target.sources:
+            key = _source_key(source)
+            if key in seen:
+                continue
+            sources.append(source)
+            seen.add(key)
+    return target._replace(sources=tuple(sources))
+
+
 def _resolve_file_target(file_path: str) -> SessionLogTarget:
     resolved = Path(file_path).expanduser().resolve()
     if not resolved.is_file():
@@ -84,11 +122,14 @@ def _resolve_file_target(file_path: str) -> SessionLogTarget:
     elif ".codex" in parts:
         harness = "codex"
 
-    return SessionLogTarget(
-        session_id=_extract_session_id_from_path(resolved),
-        harness=harness,
-        file_path=resolved,
-        source="file",
+    return _target_from_source(
+        TranscriptSource(
+            kind="file",
+            session_id=_extract_session_id_from_path(resolved),
+            harness=harness,
+            path=resolved,
+            source_label="file",
+        )
     )
 
 
@@ -105,26 +146,25 @@ def _resolve_adapter_file_target(
     )
     if candidate is None or not candidate.is_file():
         return None
-    return SessionLogTarget(
-        session_id=session_id,
-        harness=str(harness_id),
-        file_path=candidate,
-        source=f"{harness_id} transcript",
+    return _target_from_source(
+        TranscriptSource(
+            kind="file",
+            session_id=session_id,
+            harness=str(harness_id),
+            path=candidate,
+            source_label=f"{harness_id} transcript",
+        )
     )
 
 
-def _opencode_db_target(
-    *,
-    session_id: str,
-    fallback_targets: tuple[SessionLogTarget, ...] = (),
-) -> SessionLogTarget:
-    return SessionLogTarget(
-        session_id=session_id,
-        harness=HarnessId.OPENCODE.value,
-        file_path=None,
-        source="opencode transcript",
-        transcript_source="opencode_db",
-        fallback_targets=fallback_targets,
+def _opencode_db_target(*, session_id: str) -> SessionLogTarget:
+    return _target_from_source(
+        TranscriptSource(
+            kind="opencode_db",
+            session_id=session_id,
+            harness=HarnessId.OPENCODE.value,
+            source_label="opencode transcript",
+        )
     )
 
 
@@ -160,9 +200,9 @@ def _resolve_harness_session_file(
             harness_id == HarnessId.OPENCODE
             and opencode_db_session_exists(session_id=normalized_session_id)
         ):
-            return _opencode_db_target(
-                session_id=normalized_session_id,
-                fallback_targets=(file_target,) if file_target is not None else (),
+            return _with_sources(
+                _opencode_db_target(session_id=normalized_session_id),
+                file_target,
             )
         if file_target is not None:
             return file_target
@@ -187,9 +227,9 @@ def _resolve_harness_session_file(
             harness_id == HarnessId.OPENCODE
             and opencode_db_session_exists(session_id=normalized_session_id)
         ):
-            return _opencode_db_target(
-                session_id=normalized_session_id,
-                fallback_targets=(file_target,) if file_target is not None else (),
+            return _with_sources(
+                _opencode_db_target(session_id=normalized_session_id),
+                file_target,
             )
         if file_target is not None:
             return file_target
@@ -314,47 +354,15 @@ def _target_from_spawn_output(
     output_path = spawn_output_path_for_target(runtime_root, spawn_id, live_first=live_first)
     if output_path is None:
         return None
-    return SessionLogTarget(
-        session_id=display_id,
-        harness=None,
-        file_path=output_path,
-        source=source or f"spawn {spawn_id} output",
+    return _target_from_source(
+        TranscriptSource(
+            kind="spawn_history",
+            session_id=display_id,
+            harness=None,
+            path=output_path,
+            source_label=source or f"spawn {spawn_id} output",
+        )
     )
-
-
-def _target_key(target: SessionLogTarget) -> tuple[str, str, str | None, str]:
-    return (
-        target.transcript_source,
-        target.session_id,
-        target.file_path.as_posix() if target.file_path is not None else None,
-        target.source,
-    )
-
-
-def _with_fallback_targets(
-    target: SessionLogTarget,
-    *fallback_targets: SessionLogTarget | None,
-) -> SessionLogTarget:
-    existing = list(target.fallback_targets)
-    seen = {_target_key(target), *(_target_key(fallback) for fallback in existing)}
-    for fallback in fallback_targets:
-        if fallback is None:
-            continue
-        key = _target_key(fallback)
-        if key in seen:
-            continue
-        existing.append(fallback)
-        seen.add(key)
-    return target._replace(fallback_targets=tuple(existing))
-
-
-def _with_opencode_fallback_targets(
-    target: SessionLogTarget,
-    *fallback_targets: SessionLogTarget | None,
-) -> SessionLogTarget:
-    if target.transcript_source != "opencode_db":
-        return target
-    return _with_fallback_targets(target, *fallback_targets)
 
 
 def _managed_primary_output_target(
@@ -619,39 +627,6 @@ def _skip_primary_default_root_detection(
     return configured_session_dir != default_session_dir
 
 
-def _repair_target_from_detected_session_id(
-    *,
-    project_root: Path,
-    harness: str | None,
-    detected_session_id: str,
-    invalid_reason: str,
-    unavailable_reason: str,
-) -> SessionRepairTarget:
-    if _is_spawn_ref(detected_session_id):
-        return SessionRepairTarget(
-            detected_harness_session_id=None,
-            source=None,
-            reason=invalid_reason,
-        )
-
-    transcript_target = _resolve_harness_transcript_target_or_none(
-        project_root=project_root,
-        session_id=detected_session_id,
-        harness=harness,
-    )
-    if transcript_target is not None:
-        return SessionRepairTarget(
-            detected_harness_session_id=transcript_target.session_id,
-            source=transcript_target.source,
-        )
-
-    return SessionRepairTarget(
-        detected_harness_session_id=detected_session_id,
-        source=f"{harness or 'primary'} detected session id",
-        reason=unavailable_reason,
-    )
-
-
 def _resolve_from_chat_id(
     *,
     project_root: Path,
@@ -711,7 +686,7 @@ def _resolve_from_chat_id(
             chat_id=chat_id,
             primary_spawn=primary_spawn,
         )
-        return _with_opencode_fallback_targets(transcript_target, output_target)
+        return _with_sources(transcript_target, output_target)
 
     detected_session_id = _detect_primary_session_id(
         project_root=project_root,
@@ -732,7 +707,7 @@ def _resolve_from_chat_id(
                 chat_id=chat_id,
                 primary_spawn=primary_spawn,
             )
-            return _with_opencode_fallback_targets(transcript_target, output_target)
+            return _with_sources(transcript_target, output_target)
 
     if primary_spawn is not None:
         output_target = _managed_primary_output_target(
@@ -877,7 +852,7 @@ def _resolve_from_spawn_id(
             is_managed_backend_primary=is_managed_backend_primary,
             harness=harness,
         )
-        return _with_opencode_fallback_targets(transcript_target, output_target)
+        return _with_sources(transcript_target, output_target)
 
     if is_primary_spawn:
         detected_session_id = _detect_primary_session_id(
@@ -901,7 +876,7 @@ def _resolve_from_spawn_id(
                     is_managed_backend_primary=is_managed_backend_primary,
                     harness=harness,
                 )
-                return _with_opencode_fallback_targets(transcript_target, output_target)
+                return _with_sources(transcript_target, output_target)
 
     if is_primary_spawn:
         if is_managed_backend_primary:
@@ -955,7 +930,7 @@ def _resolve_from_session_ref(
             display_id=session_id,
             record=record,
         )
-        return _with_opencode_fallback_targets(target, output_target)
+        return _with_sources(target, output_target)
 
     inferred = infer_harness_from_untracked_session_ref(project_root, session_ref)
     inferred_name = str(inferred) if inferred is not None else None
@@ -969,212 +944,7 @@ def _resolve_from_session_ref(
         display_id=session_ref,
         harness_session_id=session_ref,
     )
-    return _with_opencode_fallback_targets(target, output_target)
-
-
-def _resolve_repair_from_chat_id(
-    *,
-    project_root: Path,
-    runtime_root: Path,
-    chat_id: str,
-) -> SessionRepairTarget:
-    session_record = _read_chat_session_record(runtime_root, chat_id)
-    if session_record is None:
-        raise ValueError(f"Chat '{chat_id}' not found")
-
-    primary_spawn = _primary_spawn_for_chat(project_root, runtime_root, chat_id)
-    harness = session_record.harness.strip() or None
-    if harness is None and primary_spawn is not None:
-        harness = (primary_spawn.harness or "").strip() or None
-
-    transcript_target = _resolve_transcript_from_candidates(
-        project_root=project_root,
-        harness=harness,
-        candidate_ids=[
-            _latest_harness_session_id(session_record),
-            (
-                read_primary_harness_session_id(runtime_root, primary_spawn.id)
-                if primary_spawn is not None
-                else None
-            ),
-        ],
-    )
-    if transcript_target is not None:
-        return SessionRepairTarget(
-            detected_harness_session_id=transcript_target.session_id,
-            source=transcript_target.source,
-        )
-
-    detected_session_id = _detect_primary_session_id(
-        project_root=project_root,
-        runtime_root=runtime_root,
-        spawn_row=primary_spawn,
-        harness=harness,
-    )
-    if detected_session_id is not None:
-        return _repair_target_from_detected_session_id(
-            project_root=project_root,
-            harness=harness,
-            detected_session_id=detected_session_id,
-            invalid_reason=(
-                f"Detected session id '{detected_session_id}' for chat '{chat_id}' "
-                "looks like a spawn id; refusing repair."
-            ),
-            unavailable_reason=(
-                f"Detected harness session id '{detected_session_id}' for chat '{chat_id}' "
-                "but transcript file is not available yet."
-            ),
-        )
-
-    return SessionRepairTarget(
-        detected_harness_session_id=None,
-        source=None,
-        reason=_primary_transcript_unavailable_message(chat_id),
-    )
-
-
-def _resolve_repair_from_spawn_id(
-    *,
-    project_root: Path,
-    runtime_root: Path,
-    spawn_id: str,
-) -> SessionRepairTarget:
-    row = read_spawn_row_read_only(project_root, spawn_id, runtime_root=runtime_root)
-    if row is None:
-        raise ValueError(f"Spawn '{spawn_id}' not found")
-
-    harness = (row.harness or "").strip() or None
-    row_session_id = (row.harness_session_id or "").strip()
-    transcript_target = _resolve_transcript_from_candidates(
-        project_root=project_root,
-        harness=harness,
-        candidate_ids=[row_session_id if not _is_spawn_ref(row_session_id) else None],
-    )
-    if transcript_target is not None:
-        return SessionRepairTarget(
-            detected_harness_session_id=transcript_target.session_id,
-            source=transcript_target.source,
-        )
-
-    if row.kind != "primary":
-        return SessionRepairTarget(
-            detected_harness_session_id=None,
-            source=None,
-            reason=(
-                f"Spawn '{spawn_id}' has no detected harness session id to repair "
-                "(non-primary spawns may only have output fallback)."
-            ),
-        )
-
-    if row.chat_id:
-        chat_record = _read_chat_session_record(runtime_root, row.chat_id)
-        if chat_record is not None:
-            if harness is None:
-                harness = chat_record.harness.strip() or None
-            transcript_target = _resolve_transcript_from_candidates(
-                project_root=project_root,
-                harness=harness,
-                candidate_ids=[_latest_harness_session_id(chat_record)],
-            )
-            if transcript_target is not None:
-                return SessionRepairTarget(
-                    detected_harness_session_id=transcript_target.session_id,
-                    source=transcript_target.source,
-                )
-
-    transcript_target = _resolve_transcript_from_candidates(
-        project_root=project_root,
-        harness=harness,
-        candidate_ids=[
-            (
-                read_primary_harness_session_id(runtime_root, spawn_id)
-                if row.kind == "primary"
-                else None
-            )
-        ],
-    )
-    if transcript_target is not None:
-        return SessionRepairTarget(
-            detected_harness_session_id=transcript_target.session_id,
-            source=transcript_target.source,
-        )
-
-    detected_session_id = _detect_primary_session_id(
-        project_root=project_root,
-        runtime_root=runtime_root,
-        spawn_row=row,
-        harness=harness,
-    )
-    if detected_session_id is None:
-        return SessionRepairTarget(
-            detected_harness_session_id=None,
-            source=None,
-            reason=(
-                f"Spawn '{spawn_id}' has no detected harness session id to repair "
-                "(no harness transcript available yet)."
-            ),
-        )
-
-    return _repair_target_from_detected_session_id(
-        project_root=project_root,
-        harness=harness,
-        detected_session_id=detected_session_id,
-        invalid_reason=(
-            f"Detected session id '{detected_session_id}' for spawn "
-            f"'{spawn_id}' looks like a spawn id; refusing repair."
-        ),
-        unavailable_reason=(
-            f"Detected harness session id '{detected_session_id}' for "
-            f"spawn '{spawn_id}' but transcript file is not available yet."
-        ),
-    )
-
-
-def _resolve_repair_from_session_ref(
-    *,
-    project_root: Path,
-    session_ref: str,
-) -> SessionRepairTarget:
-    inferred = infer_harness_from_untracked_session_ref(project_root, session_ref)
-    harness = str(inferred) if inferred is not None else None
-    transcript_target = _resolve_harness_transcript_target_or_none(
-        project_root=project_root,
-        session_id=session_ref,
-        harness=harness,
-    )
-    if transcript_target is not None:
-        return SessionRepairTarget(
-            detected_harness_session_id=transcript_target.session_id,
-            source=transcript_target.source,
-        )
-    raise FileNotFoundError(f"Session file for '{session_ref}' not found")
-
-
-def resolve_session_repair_target(
-    *,
-    ref: str,
-    project_root: Path,
-    runtime_root: Path,
-) -> SessionRepairTarget:
-    normalized_ref = ref.strip()
-    if not normalized_ref:
-        raise ValueError("Session reference is required")
-    if _is_chat_ref(runtime_root, normalized_ref):
-        return _resolve_repair_from_chat_id(
-            project_root=project_root,
-            runtime_root=runtime_root,
-            chat_id=normalized_ref,
-        )
-    if _is_spawn_ref(normalized_ref):
-        return _resolve_repair_from_spawn_id(
-            project_root=project_root,
-            runtime_root=runtime_root,
-            spawn_id=normalized_ref,
-        )
-    return _resolve_repair_from_session_ref(
-        project_root=project_root,
-        session_ref=normalized_ref,
-    )
+    return _with_sources(target, output_target)
 
 
 def resolve_session_log_target(
@@ -1214,8 +984,6 @@ def resolve_session_log_target(
 
 __all__ = [
     "SessionLogTarget",
-    "SessionRepairTarget",
     "resolve_session_log_target",
-    "resolve_session_repair_target",
     "spawn_output_path_for_target",
 ]
